requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
attrs==23.2.0
auth0-python==4.7.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
frozenlist==1.4.1
idna==3.7
macholib==1.16.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
setuptools==70.3.0
tabulate==0.9.0
urllib3==2.2.2
validators==0.32.0
yarl==1.9.4